@@ -10,59 +10,51 @@
     "test": "mocha --require @babel/register test/**/*.js"
   },
   "dependencies": {
-    "@babel/cli": "^7.22.10",
-    "@babel/core": "^7.22.10",
-    "@babel/node": "^7.22.10",
-    "@babel/plugin-transform-runtime": "^7.22.10",
-    "@babel/preset-env": "^7.22.10",
-    "@babel/register": "^7.22.5",
-    "@babel/runtime": "^7.22.10",
     "@popperjs/core": "^2.11.8",
-    "babel-loader": "^9.1.3",
+    "@types/cookie-parser": "^1.4.3",
+    "@types/jquery": "^3.5.17",
+    "@types/nunjucks": "^3.2.3",
+    "@types/serve-favicon": "^2.5.4",
+    "animate.css": "^4.1.1",
+    "axios": "^1.5.0",
     "body-parser": "^1.20.2",
+    "bootstrap": "^5.3.1",
     "cookie-parser": "^1.4.6",
+    "dayjs": "^1.11.9",
     "debug": "^4.3.4",
     "express": "^4.18.2",
     "express-minify-html": "^0.12.0",
     "express-sanitizer": "^1.0.6",
     "express-validator": "^7.0.1",
+    "jquery": "^3.7.1",
     "log4js": "^6.9.1",
-    "nodemon": "^3.0.1",
     "nunjucks": "^3.2.4",
     "serve-favicon": "^2.5.0"
   },
   "devDependencies": {
-    "animate.css": "^4.1.1",
-    "axios": "^1.4.0",
-    "bootstrap": "^5.3.1",
+    "@types/express": "^4.17.17",
+    "@types/mocha": "^10.0.1",
+    "@types/node": "^20.5.7",
     "clean-webpack-plugin": "^4.0.0",
     "copy-webpack-plugin": "^11.0.0",
     "css-loader": "^6.8.1",
     "cssnano": "^6.0.1",
-    "eslint": "^8.47.0",
+    "eslint": "^8.48.0",
     "eslint-plugin-simple-import-sort": "^10.0.0",
     "express-generator": "^4.16.1",
     "glob": "^10.3.3",
-    "jquery": "^3.7.0",
     "less": "^4.2.0",
     "less-loader": "^11.1.3",
-    "lodash": "^4.17.21",
     "mini-css-extract-plugin": "^2.7.6",
     "mocha": "^10.2.0",
-    "moment": "^2.29.4",
+    "nodemon": "^3.0.1",
     "postcss-safe-parser": "^6.0.0",
     "style-loader": "^3.3.3",
+    "ts-loader": "^9.4.4",
+    "typescript": "^5.2.2",
     "url-loader": "^4.1.1",
     "webpack": "^5.88.2",
     "webpack-assets-manifest": "^5.1.0",
     "webpack-cli": "^5.1.4"
-  },
-  "babel": {
-    "presets": [
-      "@babel/env"
-    ],
-    "plugins": [
-      "@babel/transform-runtime"
-    ]
   }
 }
