@@ -1,17 +1,13 @@
-const controller = require("../controller");
+import controller from "../controller";
 
-const routes = {
+export const routes = {
   "/routing": controller.routing,
 };
 
-const menu = [
+export const menu = [
   { text: "Routing", url: "/routing" },
   { text: "Middleware", url: "/middleware" },
   { text: "Error Handing", url: "/error-handing" },
   { text: "Debug", url: "/debug" },
   { text: "Database", url: "/database" },
 ];
-
-module.exports = {
-  routes, menu
-};
