@@ -1,5 +1,5 @@
-import $ from "jquery";
 import "bootstrap";
+import $ from "jquery";
 
 import ns from "../common/ns";
 
